Huge Changes (Added Major Things)
The first thing I added here was the new sprites and animations, the portal and the bullet along with the new poison floor all got animated sprites.

The poison floor along with the walls and floor are all made from the tileset, which I added.

I also got the portal working and addeda few lines of code explenation